service-workers/service-worker/clients-get-client-types.https.html

<!DOCTYPE html>
<title>Service Worker: Clients.get with window and worker clients</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
const SCOPE = 'resources/clients-get-client-types';
const SCRIPT = 'resources/clients-get-worker.js';
const FRAME_URL = SCOPE + '-frame.html';
const SHARED_WORKER_URL = SCOPE + '-shared-worker.js';
const WORKER_URL = SCOPE + '-worker.js';
let client_ids = [];

promise_test(async t => {
  t.add_cleanup(async () => {
    if (frame) frame.remove();
    if (registration) await registration.unregister();
  });

  const registration = await service_worker_unregister_and_register(
    t,
    SCRIPT,
    SCOPE
  );
  await wait_for_state(t, registration.installing, 'activated');
  const frame = await with_iframe(FRAME_URL);
  frame.focus();
  let client_id = await wait_for_clientId();
  client_ids.push(client_id);

  client_id = await new Promise(function(resolve) {
    const sharedWorker = new SharedWorker(SHARED_WORKER_URL);
    sharedWorker.port.onmessage = function(event) {
      resolve(event.data.clientId);
    };
  });
  client_ids.push(client_id);

  let channel = new MessageChannel();
  const worker = new Worker(WORKER_URL);
  worker.postMessage({ cmd: 'GetClientId', port: channel.port2 }, [
    channel.port2
  ]);
  client_id = await new Promise(function(resolve) {
    channel.port1.onmessage = function(event) {
      resolve(event.data.clientId);
    };
  });
  client_ids.push(client_id);

  channel = new MessageChannel();
  frame.contentWindow.postMessage('StartWorker', '*', [channel.port2]);
  client_id = await new Promise(function(resolve) {
    channel.port1.onmessage = function(event) {
      resolve(event.data.clientId);
    };
  });
  client_ids.push(client_id);

  const e = await new Promise(function(resolve) {
    navigator.serviceWorker.onmessage = resolve;
    registration.active.postMessage({ clientIds: client_ids });
  });
  assert_equals(e.data.length, expected.length);
  // We use these assert_not_equals because assert_array_equals doesn't
  // print the error description when passed an undefined value.
  assert_not_equals(e.data[0], undefined,
    'Window client should not be undefined');
  assert_array_equals(e.data[0], expected[0], 'Window client');
  assert_not_equals(e.data[1], undefined,
    'Shared worker client should not be undefined');
  assert_array_equals(e.data[1], expected[1], 'Shared worker client');
  assert_not_equals(e.data[2], undefined,
    'Worker(Started by main frame) client should not be undefined');
  assert_array_equals(e.data[2], expected[2],
    'Worker(Started by main frame) client');
  assert_not_equals(e.data[3], undefined,
    'Worker(Started by sub frame) client should not be undefined');
  assert_array_equals(e.data[3], expected[3],
    'Worker(Started by sub frame) client');
});

function wait_for_clientId() {
  return new Promise(function(resolve) {
    function get_client_id(e) {
      window.removeEventListener('message', get_client_id);
      resolve(e.data.clientId);
    }
    window.addEventListener('message', get_client_id, false);
  });
}

const expected = [
  // visibilityState, focused, url, type, frameType
  ['visible', true, normalizeURL(SCOPE) + '-frame.html', 'window', 'nested'],
  [undefined, undefined, normalizeURL(SCOPE) + '-shared-worker.js', 'sharedworker', 'none'],
  [undefined, undefined, normalizeURL(SCOPE) + '-worker.js', 'worker', 'none'],
  [undefined, undefined, normalizeURL(SCOPE) + '-frame-worker.js', 'worker', 'none']
];
</script>
