third_party/blink/web_tests/virtual/omt-worker-fetch/external/wpt/service-workers/service-worker/clients-get-client-types.https-expected.txt

This is a testharness.js-based test.
FAIL Test Clients.get() with window and worker clients assert_array_equals: Worker(Started by main frame) client property 3, expected "worker" but got "sharedworker"
Harness: the test ran to completion.
